refactor: consolidate rpc transports
