--- src/lib/rpc/message-port.ts
@@ -1,9 +1,0 @@
-import type { RpcClient } from "./core.ts";
-import { createEndpointRpc } from "./endpoint.ts";
-
-export function createMessagePortRpc<Handlers>(
-  port: MessagePort,
-): RpcClient<Handlers> {
-  port.start();
-  return createEndpointRpc<Handlers>(port);
-}
